More cleaning ;-)
git-svn-id: https://svn.code.sf.net/p/fsthost/code/trunk@196 75a15a69-0cfc-47d9-a1bd-ce2941b697b2

+++ jackvst.h
@@ -47,8 +47,8 @@ struct _JackVST {
     FST*            fst;
     char*           client_name;
     char*           default_state_file;
-    short           numIns;
-    short           numOuts;
+    int32_t         numIns;
+    int32_t         numOuts;
     float**         ins;
     float**         outs;
     jack_port_t*    midi_inport;
